+++ src/core/gdbstub/gdbstub.cpp
@@ -302,7 +302,9 @@ static void RemoveBreakpoint(BreakpointType type, PAddr addr) {
     auto bp = p.find(addr);
     if (bp != p.end()) {
         LOG_DEBUG(Debug_GDBStub, "gdb: removed a breakpoint: %08x bytes at %08x of type %d\n",
-                  bp->second.len, bp->second.addr, type);
+                  static_cast<u32>(bp->second.len),
+                  static_cast<u32>(bp->second.addr),
+                  static_cast<int>(type));
         p.erase(addr);
     }
 }
@@ -348,7 +350,8 @@ bool CheckBreakpoint(PAddr addr, BreakpointType type) {
 
         if (bp->second.active && (addr >= bp->second.addr && addr < bp->second.addr + len)) {
             LOG_DEBUG(Debug_GDBStub,
-                      "Found breakpoint type %d @ %08x, range: %08x - %08x (%d bytes)\n", type,
+                      "Found breakpoint type %d @ %08x, range: %08x - %08x (%d bytes)\n",
+                      static_cast<int>(type),
                       addr, bp->second.addr, bp->second.addr + len, len);
             return true;
         }
@@ -734,7 +737,8 @@ static bool CommitBreakpoint(BreakpointType type, PAddr addr, u32 len) {
     breakpoint.len = len;
     p.insert({addr, breakpoint});
 
-    LOG_DEBUG(Debug_GDBStub, "gdb: added %d breakpoint: %08x bytes at %08x\n", type, breakpoint.len,
+    LOG_DEBUG(Debug_GDBStub, "gdb: added %d breakpoint: %08x bytes at %08x\n",
+              static_cast<int>(type), breakpoint.len,
               breakpoint.addr);
 
     return true;
