+++ src/core/hle/service/ndm/ndm.cpp
@@ -32,7 +32,8 @@ void EnterExclusiveState(Service::Interface* self) {
 
     cmd_buff[0] = IPC::MakeHeader(0x1, 1, 0);
     cmd_buff[1] = RESULT_SUCCESS.raw; // No error
-    LOG_WARNING(Service_NDM, "(STUBBED) exclusive_state=0x%08X ", exclusive_state);
+    LOG_WARNING(Service_NDM, "(STUBBED) exclusive_state=0x%08X ",
+                static_cast<u32>(exclusive_state));
 }
 
 void LeaveExclusiveState(Service::Interface* self) {
@@ -41,7 +42,8 @@ void LeaveExclusiveState(Service::Interface* self) {
 
     cmd_buff[0] = IPC::MakeHeader(0x2, 1, 0);
     cmd_buff[1] = RESULT_SUCCESS.raw; // No error
-    LOG_WARNING(Service_NDM, "(STUBBED) exclusive_state=0x%08X ", exclusive_state);
+    LOG_WARNING(Service_NDM, "(STUBBED) exclusive_state=0x%08X ",
+                static_cast<u32>(exclusive_state));
 }
 
 void QueryExclusiveMode(Service::Interface* self) {
@@ -50,7 +52,8 @@ void QueryExclusiveMode(Service::Interface* self) {
     cmd_buff[0] = IPC::MakeHeader(0x3, 2, 0);
     cmd_buff[1] = RESULT_SUCCESS.raw; // No error
     cmd_buff[2] = static_cast<u32>(exclusive_state);
-    LOG_WARNING(Service_NDM, "(STUBBED) exclusive_state=0x%08X ", exclusive_state);
+    LOG_WARNING(Service_NDM, "(STUBBED) exclusive_state=0x%08X ",
+                static_cast<u32>(exclusive_state));
 }
 
 void LockState(Service::Interface* self) {
@@ -84,7 +87,8 @@ void SuspendDaemons(Service::Interface* self) {
 
     cmd_buff[0] = IPC::MakeHeader(0x6, 1, 0);
     cmd_buff[1] = RESULT_SUCCESS.raw; // No error
-    LOG_WARNING(Service_NDM, "(STUBBED) daemon_bit_mask=0x%08X ", daemon_bit_mask);
+    LOG_WARNING(Service_NDM, "(STUBBED) daemon_bit_mask=0x%08X ",
+                static_cast<u32>(daemon_bit_mask));
 }
 
 void ResumeDaemons(Service::Interface* self) {
@@ -99,7 +103,8 @@ void ResumeDaemons(Service::Interface* self) {
 
     cmd_buff[0] = IPC::MakeHeader(0x7, 1, 0);
     cmd_buff[1] = RESULT_SUCCESS.raw; // No error
-    LOG_WARNING(Service_NDM, "(STUBBED) daemon_bit_mask=0x%08X ", daemon_bit_mask);
+    LOG_WARNING(Service_NDM, "(STUBBED) daemon_bit_mask=0x%08X ",
+                static_cast<u32>(daemon_bit_mask));
 }
 
 void SuspendScheduler(Service::Interface* self) {
@@ -198,7 +203,8 @@ void OverrideDefaultDaemons(Service::Interface* self) {
 
     cmd_buff[0] = IPC::MakeHeader(0x14, 1, 0);
     cmd_buff[1] = RESULT_SUCCESS.raw; // No error
-    LOG_WARNING(Service_NDM, "(STUBBED) default_daemon_bit_mask=0x%08X ", default_daemon_bit_mask);
+    LOG_WARNING(Service_NDM, "(STUBBED) default_daemon_bit_mask=0x%08X ",
+                static_cast<u32>(default_daemon_bit_mask));
 }
 
 void ResetDefaultDaemons(Service::Interface* self) {
@@ -207,7 +213,8 @@ void ResetDefaultDaemons(Service::Interface* self) {
 
     cmd_buff[0] = IPC::MakeHeader(0x15, 1, 0);
     cmd_buff[1] = RESULT_SUCCESS.raw; // No error
-    LOG_WARNING(Service_NDM, "(STUBBED) default_daemon_bit_mask=0x%08X ", default_daemon_bit_mask);
+    LOG_WARNING(Service_NDM, "(STUBBED) default_daemon_bit_mask=0x%08X ",
+                static_cast<u32>(default_daemon_bit_mask));
 }
 
 void GetDefaultDaemons(Service::Interface* self) {
@@ -216,7 +223,8 @@ void GetDefaultDaemons(Service::Interface* self) {
     cmd_buff[0] = IPC::MakeHeader(0x16, 2, 0);
     cmd_buff[1] = RESULT_SUCCESS.raw; // No error
     cmd_buff[2] = static_cast<u32>(default_daemon_bit_mask);
-    LOG_WARNING(Service_NDM, "(STUBBED) default_daemon_bit_mask=0x%08X ", default_daemon_bit_mask);
+    LOG_WARNING(Service_NDM, "(STUBBED) default_daemon_bit_mask=0x%08X ",
+                static_cast<u32>(default_daemon_bit_mask));
 }
 
 void ClearHalfAwakeMacFilter(Service::Interface* self) {
