+++ src/video_core/renderer_opengl/pica_to_gl.h
@@ -89,7 +89,7 @@ inline GLenum BlendEquation(Pica::FramebufferRegs::BlendEquation equation) {
 
     // Range check table for input
     if (static_cast<size_t>(equation) >= ARRAY_SIZE(blend_equation_table)) {
-        LOG_CRITICAL(Render_OpenGL, "Unknown blend equation %d", equation);
+        LOG_CRITICAL(Render_OpenGL, "Unknown blend equation %d", static_cast<int>(equation));
         UNREACHABLE();
 
         return GL_FUNC_ADD;
@@ -119,7 +119,7 @@ inline GLenum BlendFunc(Pica::FramebufferRegs::BlendFactor factor) {
 
     // Range check table for input
     if (static_cast<size_t>(factor) >= ARRAY_SIZE(blend_func_table)) {
-        LOG_CRITICAL(Render_OpenGL, "Unknown blend factor %d", factor);
+        LOG_CRITICAL(Render_OpenGL, "Unknown blend factor %d", static_cast<int>(factor));
         UNREACHABLE();
 
         return GL_ONE;
@@ -150,7 +150,7 @@ inline GLenum LogicOp(Pica::FramebufferRegs::LogicOp op) {
 
     // Range check table for input
     if (static_cast<size_t>(op) >= ARRAY_SIZE(logic_op_table)) {
-        LOG_CRITICAL(Render_OpenGL, "Unknown logic op %d", op);
+        LOG_CRITICAL(Render_OpenGL, "Unknown logic op %d", static_cast<int>(op));
         UNREACHABLE();
 
         return GL_COPY;
@@ -173,7 +173,7 @@ inline GLenum CompareFunc(Pica::FramebufferRegs::CompareFunc func) {
 
     // Range check table for input
     if (static_cast<size_t>(func) >= ARRAY_SIZE(compare_func_table)) {
-        LOG_CRITICAL(Render_OpenGL, "Unknown compare function %d", func);
+        LOG_CRITICAL(Render_OpenGL, "Unknown compare function %d", static_cast<int>(func));
         UNREACHABLE();
 
         return GL_ALWAYS;
@@ -196,7 +196,7 @@ inline GLenum StencilOp(Pica::FramebufferRegs::StencilAction action) {
 
     // Range check table for input
     if (static_cast<size_t>(action) >= ARRAY_SIZE(stencil_op_table)) {
-        LOG_CRITICAL(Render_OpenGL, "Unknown stencil op %d", action);
+        LOG_CRITICAL(Render_OpenGL, "Unknown stencil op %d", static_cast<int>(action));
         UNREACHABLE();
 
         return GL_KEEP;
